fix: API call and display issues (#277)
* LAF item view moved to top level to match other detail page paths.
* Detail content structure adjusted to match header display and gutter sizes, removed superfluous status display (already shown at the top).
* LAF query renamed to match other queries.
* LAF item API query removed (I don't see anything in the API spec for it).
* Stack screen string adjusted for LAF detail page.
